update Solidity types to use int256 for compatibility; add EVM version settings and new scripts for bytecode verification and Java wrapper updates

+++ src/main/resources/contracts/IProperty.sol
@@ -2,8 +2,8 @@
 pragma solidity ^0.8.0;
 
 interface IProperty {
-    function setValue(int _value) external;
-    function getValue() external view returns (int);
+    function setValue(int256 _value) external;
+    function getValue() external view returns (int256);
     function setPropertyId(string memory _propertyId) external;
     function getPropertyId() external view returns (string memory);
     function setName(string memory _name) external;
